SEP6-28: added interface and appropriate error handling
